Fixed VLArray to host more than 32-bit long rows

+++ src/H5VLARRAY.c
@@ -197,7 +197,7 @@ out:
 
 herr_t H5VLARRAYappend_records( hid_t dataset_id,
                                 hid_t type_id,
-                                int nobjects,
+                                size_t nobjects,
                                 hsize_t nrecords,
                                 const void *data )
 {
@@ -275,7 +275,7 @@ out:
 herr_t H5VLARRAYmodify_records( hid_t dataset_id,
                                 hid_t type_id,
                                 hsize_t nrow,
-                                int nobjects,
+                                size_t nobjects,
                                 const void *data )
 {
